requirements.txt
Django==5.0.1
cryptography==42.0.1
PyJWT==2.8.0
